pylbo: refactored interactive eigenfunction handlers
