+++ post_processing/pylbo/visualisation/spectra/spectrum_multi.py
@@ -2,8 +2,10 @@ import numpy as np
 from matplotlib import colors
 from pylbo.utilities.toolbox import add_pickradius_to_item, transform_to_numpy
 from pylbo.visualisation.continua import ContinuaHandler
-from pylbo.visualisation.derived_eigenfunctions import DerivedEigenfunctionHandler
-from pylbo.visualisation.eigenfunctions import EigenfunctionHandler
+from pylbo.visualisation.eigenfunctions.derived_eigfunc_handler import (
+    DerivedEigenfunctionHandler,
+)
+from pylbo.visualisation.eigenfunctions.eigfunc_handler import EigenfunctionHandler
 from pylbo.visualisation.spectra.spectrum_figure import SpectrumFigure
